@@ -66,89 +64,94 @@
     "url": "https://github.com/ColorlibHQ/AdminLTE/issues"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.11.2",
-    "@fullcalendar/bootstrap": "^4.3.0",
-    "@fullcalendar/core": "^4.3.1",
-    "@fullcalendar/daygrid": "^4.3.0",
-    "@fullcalendar/interaction": "^4.3.0",
-    "@fullcalendar/timegrid": "^4.3.0",
+    "@fortawesome/fontawesome-free": "^5.15.2",
     "@lgaitan/pace-progress": "^1.0.7",
-    "@sweetalert2/theme-bootstrap-4": "^2.2.1",
-    "@ttskch/select2-bootstrap4-theme": "^1.3.2",
-    "bootstrap": "^4.3.1",
-    "bootstrap-colorpicker": "^3.1.2",
-    "bootstrap-slider": "^10.6.2",
-    "bootstrap-switch": "^3.4.0",
-    "bootstrap4-duallistbox": "^4.0.1",
-    "bs-custom-file-input": "^1.3.2",
-    "chart.js": "^2.9.3",
-    "datatables.net": "^1.10.20",
-    "datatables.net-autofill-bs4": "^2.3.4",
-    "datatables.net-bs4": "^1.10.20",
-    "datatables.net-buttons-bs4": "^1.6.1",
-    "datatables.net-colreorder-bs4": "^1.5.2",
-    "datatables.net-fixedcolumns-bs4": "^3.3.0",
-    "datatables.net-fixedheader-bs4": "^3.1.6",
-    "datatables.net-keytable-bs4": "^2.5.1",
-    "datatables.net-responsive-bs4": "^2.2.3",
-    "datatables.net-rowgroup-bs4": "^1.1.1",
-    "datatables.net-rowreorder-bs4": "^1.2.6",
-    "datatables.net-scroller-bs4": "^2.0.1",
+    "@sweetalert2/theme-bootstrap-4": "^4.0.3",
+    "@ttskch/select2-bootstrap4-theme": "^1.5.2",
+    "bootstrap": "^4.6.0",
+    "bootstrap-colorpicker": "^3.2.0",
+    "bootstrap-slider": "^11.0.2",
+    "bootstrap-switch": "3.3.4",
+    "bootstrap4-duallistbox": "^4.0.2",
+    "bs-custom-file-input": "^1.3.4",
+    "bs-stepper": "^1.7.0",
+    "chart.js": "^2.9.4",
+    "codemirror": "^5.59.4",
+    "datatables.net": "^1.10.23",
+    "datatables.net-autofill-bs4": "^2.3.5",
+    "datatables.net-bs4": "^1.10.23",
+    "datatables.net-buttons-bs4": "^1.6.5",
+    "datatables.net-colreorder-bs4": "^1.5.3",
+    "datatables.net-fixedcolumns-bs4": "^3.3.2",
+    "datatables.net-fixedheader-bs4": "^3.1.8",
+    "datatables.net-keytable-bs4": "^2.6.1",
+    "datatables.net-responsive-bs4": "^2.2.7",
+    "datatables.net-rowgroup-bs4": "^1.1.2",
+    "datatables.net-rowreorder-bs4": "^1.2.7",
+    "datatables.net-scroller-bs4": "^2.0.3",
+    "datatables.net-searchbuilder-bs4": "^1.0.1",
+    "datatables.net-searchpanes-bs4": "^1.2.2",
     "datatables.net-select-bs4": "^1.3.1",
-    "daterangepicker": "^3.0.5",
+    "daterangepicker": "^3.1.0",
+    "dropzone": "^5.8.0",
     "ekko-lightbox": "^5.3.0",
     "fastclick": "^1.0.6",
-    "filterizr": "^2.2.3",
-    "flag-icon-css": "^3.4.5",
-    "flot": "^3.2.13",
-    "fs-extra": "^8.1.0",
+    "filterizr": "^2.2.4",
+    "flag-icon-css": "^3.5.0",
+    "flot": "^4.2.1",
+    "fs-extra": "^9.1.0",
+    "fullcalendar": "^5.5.1",
     "icheck-bootstrap": "^3.0.1",
-    "inputmask": "^4.0.9",
-    "ion-rangeslider": "^2.3.0",
-    "jquery": "^3.4.1",
+    "inputmask": "^5.0.5",
+    "ion-rangeslider": "^2.3.1",
+    "jquery": "^3.6.0",
     "jquery-knob-chif": "^1.2.13",
     "jquery-mapael": "^2.2.0",
     "jquery-mousewheel": "^3.1.13",
     "jquery-ui-dist": "^1.12.1",
-    "jquery-validation": "^1.19.1",
+    "jquery-validation": "^1.19.3",
     "jqvmap-novulnerability": "^1.5.1",
     "jsgrid": "^1.5.3",
-    "jszip": "^3.2.2",
-    "moment": "^2.24.0",
-    "overlayscrollbars": "^1.10.0",
-    "pdfmake": "^0.1.62",
-    "popper.js": "^1.16.0",
+    "jszip": "^3.6.0",
+    "moment": "^2.29.1",
+    "overlayscrollbars": "^1.13.1",
+    "pdfmake": "^0.1.70",
+    "popper.js": "^1.16.1",
     "raphael": "^2.3.0",
-    "select2": "^4.0.12",
-    "sparklines": "^1.2.0",
-    "summernote": "^0.8.12",
-    "sweetalert2": "^8.19.0",
-    "tempusdominus-bootstrap-4": "^5.1.2",
-    "toastr": "^2.1.4"
+    "select2": "^4.0.13",
+    "sparklines": "^1.3.0",
+    "summernote": "^0.8.18",
+    "sweetalert2": "^10.15.5",
+    "tempusdominus-bootstrap-4": "^5.39.0",
+    "toastr": "^2.1.4",
+    "uplot": "^1.6.5"
   },
   "devDependencies": {
-    "@babel/cli": "^7.7.4",
-    "@babel/core": "^7.7.4",
-    "@babel/plugin-external-helpers": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "autoprefixer": "^9.7.3",
-    "babel-eslint": "^10.0.3",
-    "browser-sync": "^2.26.7",
-    "clean-css-cli": "^4.3.0",
-    "css-loader": "^3.2.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-compat": "^3.3.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "node-sass": "^4.13.0",
+    "@babel/core": "^7.13.8",
+    "@babel/preset-env": "^7.13.9",
+    "@rollup/plugin-babel": "^5.3.0",
+    "autoprefixer": "^10.2.5",
+    "browser-sync": "^2.26.14",
+    "bundlewatch": "^0.3.2",
+    "clean-css-cli": "^5.2.1",
+    "concurrently": "^6.0.0",
+    "eslint": "^7.21.0",
+    "eslint-config-xo": "^0.35.0",
+    "eslint-plugin-compat": "^3.9.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-unicorn": "^28.0.2",
+    "lockfile-lint": "^4.6.2",
+    "node-sass": "^5.0.0",
     "node-sass-package-importer": "^5.3.2",
-    "nodemon": "^1.19.4",
+    "nodemon": "^2.0.7",
     "npm-run-all": "^4.1.5",
-    "path": "^0.12.7",
-    "postcss-cli": "^6.1.3",
-    "rollup": "^1.27.8",
-    "rollup-plugin-babel": "^4.3.3",
-    "set-value": "^3.0.1",
-    "style-loader": "^1.0.1",
-    "terser": "^4.4.2"
+    "postcss": "^8.2.7",
+    "postcss-cli": "^8.3.1",
+    "postcss-scrollbar": "^0.3.0",
+    "rollup": "^2.40.0",
+    "stylelint": "^13.12.0",
+    "stylelint-config-twbs-bootstrap": "^2.1.0",
+    "terser": "^5.6.0",
+    "vnu-jar": "^21.2.5"
   }
 }
