@@ -14,19 +14,27 @@ import { AccountComponent } from './account/account.component';
 import { OrdersComponent } from './orders/orders.component';
 import { StatisticsComponent } from './statistics/statistics.component';
 import { HttpClientModule } from '@angular/common/http';
-import { ProductsService } from '../services/products.service';
 import { BrowserModule } from '@angular/platform-browser';
+import { ProductsComponent } from './products/products.component';
+import { FormsModule } from '@angular/forms';
+import { NgxPaginationModule } from 'ngx-pagination';
+import { Ng2SearchPipeModule } from 'ng2-search-filter';
+import { Ng2OrderModule } from 'ng2-order-pipe';
 
 
 
 @NgModule({
-  declarations: [AdminComponent, SidebarComponent, TopbarComponent, FooterComponent, AsidebarComponent,LoginComponent, AccountComponent, OrdersComponent, StatisticsComponent],
+  declarations: [AdminComponent, SidebarComponent, TopbarComponent, FooterComponent, AsidebarComponent,LoginComponent, AccountComponent, OrdersComponent, StatisticsComponent, ProductsComponent],
   imports: [
     CommonModule,
     AdminRoutingModule,
-    HttpClientModule
+    HttpClientModule,
+    FormsModule,
+    NgxPaginationModule,
+    Ng2SearchPipeModule,
+    Ng2OrderModule
     
   ],  
-  providers: [ProductsService],
+  providers: []
 })
 export class AdminModule { }
