@@ -7,12 +7,13 @@ using System.Threading.Tasks;
 
 namespace API.Models
 {
-    public class AppUser : IdentityUser
+    public class AppUser : IdentityUser<int>
     {
+
         public string FirstName { get; set; }
         public string LastName { get; set; }
 
-        //public virtual ICollection<Order> Orders { get; set; }
+        public virtual ICollection<Order> Orders { get; set; }
         
     }
 }
