@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using API.Models;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace Models
 {
@@ -18,10 +20,12 @@ namespace Models
         // thiết lập quan hệ 1-n, 1 user có nhiều order
         //public string UserId { get; set; }
 
-        //public AppUser AppUser { get; set; }
-
+        public AppUser AppUser { get; set; }
+        public int User_Id {get;set;}
 
         // 1 Order co nhieu OrderDetails
+
+        [JsonIgnore]
         public virtual ICollection<OrderDetail> OrderDetails { get; set; }
     }
 }
