@@ -1,4 +1,6 @@
+import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
+import { OrderService } from 'src/app/services/order.service';
 
 @Component({
   selector: 'app-orders',
@@ -6,10 +8,38 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./orders.component.css']
 })
 export class OrdersComponent implements OnInit {
+  orders: any = this.orderService.getOrders();
+  url = 'https://localhost:5001/api';
+  constructor(
+    private orderService: OrderService,
+    private http: HttpClient
+  ) {}
 
-  constructor() { }
 
   ngOnInit(): void {
+    this.http.get(this.url + '/orders').subscribe(data => console.log(data),
+    error => console.log(error)
+    )
   }
+  search(input: any) {
+    //this.proSer.getFromDb(input);
+  }
+  onSelect(event: any) {
 
+  }
+  add(){
+
+  }
+  selectForUpdate(){
+
+  }
+  update(){
+
+  }
+  deleteConForm(){
+
+  }
+  selectForDelete(){
+
+  }
 }
