@@ -1,36 +1,47 @@
 import { Component, OnInit } from '@angular/core';
-import { productModel } from 'src/app/models/product.model';
+import { HttpClient } from '@angular/common/http';
 import { ProductsService } from 'src/app/services/products.service';
+import { productModel } from 'src/app/models/product.model';
 
-// Perry
-/* declare var $
-declare var toastr;
-declare var  Toast; */
 @Component({
   selector: 'app-products',
   templateUrl: './products.component.html',
   styleUrls: ['./products.component.css'],
 })
 export class ProductsComponent implements OnInit {
-  // Perry
-  products:productModel[] = [];
-  // SelectedImage:File=null;
 
- 
-  constructor(private service:ProductsService) {}
+  products : productModel[] = [];
+  productName : string = "";
+  p : number =1;
   
+  constructor(public productService: ProductsService){}
+
   ngOnInit(): void {
-    this.service.getProduct().subscribe(
-      (data:productModel[]) => {
-        console.log(data);
-        this.products = data;
-      }
-    )
-    
+    this.productService.getProduct().subscribe((data: productModel[]) => {this.products = data});
+  }
+  search(input: any) {
   }
 
-  search(input: any) {
-    //this.proSer.getFromDb(input);
+  Search(){
+    if(this.productName == ""){
+      this.ngOnInit();
+    }
+    else
+    {
+      this.products = this.products.filter((res) =>
+        {
+          return res.name.toLowerCase().match(this.productName.toLowerCase());
+          
+        });
+    }
+  }
+
+  key: string = 'id';
+  reverse: boolean = false;
+  Sort(key: string)
+  {
+    this.key = key;
+    this.reverse = !this.reverse;
   }
   onSelect(event: any) {
 
@@ -50,5 +61,7 @@ export class ProductsComponent implements OnInit {
   selectForDelete(){
 
   }
+
+  
   
 }
