@@ -2,8 +2,8 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-export const environment = {
-  production: false,
+export const environment = {  
+  production: true,
   urlAPI:"https://localhost:5001/api"
 };
 
